Automatically synchronize CoreInfo when on connect and disconnect
This refactors CoreInfo to be a single common class rather than
separate classes for the core and client, adjusts it to properly
use the SyncableObject infrastructure, and adds support for
dynamically updating the list on both the client and core when a
client connects or disconnects, keeping the list up-to-date at all
times.  Forwards and backwards compatibility is maintained, though
the updated client is required in order for the automatic updating
to work.

Signed-off-by: Janne Koschinski <janne@kuschku.de>

Co-authored-by: Janne Koschinski <janne@kuschku.de>
Co-authored-by: Michael Marley <michael@michaelmarley.com>